MINIFICPP-1698 - Move provider impl, preconditions for read/write
